feat: update input implementation
