+++ src/ui/core/input.tsx
@@ -1,11 +1,18 @@
 import * as React from 'react';
+import type {
+  Control,
+  FieldValues,
+  Path,
+  RegisterOptions,
+} from 'react-hook-form';
+import { useController } from 'react-hook-form';
 import type { TextInput, TextInputProps } from 'react-native';
 import { I18nManager, StyleSheet, View } from 'react-native';
 import { TextInput as NTextInput } from 'react-native';
 import { tv } from 'tailwind-variants';
 
-import colors from '../../theme/colors';
-import { Text } from '../text';
+import colors from '../theme/colors';
+import { Text } from './text';
 
 const inputTv = tv({
   slots: {
@@ -45,6 +52,22 @@ export interface NInputProps extends TextInputProps {
   disabled?: boolean;
   error?: string;
 }
+
+type TRule = Omit<
+  RegisterOptions,
+  'valueAsNumber' | 'valueAsDate' | 'setValueAs'
+>;
+
+export type RuleType<T> = { [name in keyof T]: TRule };
+export type InputControllerType<T extends FieldValues> = {
+  name: Path<T>;
+  control: Control<T>;
+  rules?: TRule;
+};
+
+interface ControlledInputProps<T extends FieldValues>
+  extends NInputProps,
+    InputControllerType<T> {}
 
 export const Input = React.forwardRef<TextInput, NInputProps>((props, ref) => {
   const { label, error, ...inputProps } = props;
@@ -86,3 +109,22 @@ export const Input = React.forwardRef<TextInput, NInputProps>((props, ref) => {
     </View>
   );
 });
+
+// only used with react-hook-form
+export function ControlledInput<T extends FieldValues>(
+  props: ControlledInputProps<T>
+) {
+  const { name, control, rules, ...inputProps } = props;
+
+  const { field, fieldState } = useController({ control, name, rules });
+  return (
+    <Input
+      ref={field.ref}
+      autoCapitalize="none"
+      onChangeText={field.onChange}
+      value={field.value as string}
+      {...inputProps}
+      error={fieldState.error?.message}
+    />
+  );
+}
