--- src/ui/core/input/index.tsx
@@ -1,2 +1,0 @@
-export * from './controlled-input';
-export * from './input';
